Moving some functionality to workers, trying different wallets
